Update patch set 2
Patch Set 2:

I tried compiling the updated version but it just fails with the following error:

> src/third_party/skia at 328ca72d8
> You have unstaged changes.
> Please commit, stash, or reset.

Then it starts to clone the lineage channel. Another thing I've noticed is that the script expects a env/swe/channels folder, but the command is executed before it finishes fetching the source code.

Patch-set: 2